# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Eventos",
        "description": "Documentação da API de Eventos",
        "contact": {
            "email": "dwerlich21@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Obter lista de eventos",
                "description": "Retorna todos os eventos com os usuários associados.",
                "operationId": "b2e55feb6b984fe8bede8d15c4028b7d",
                "responses": {
                    "200": {
                        "description": "Lista de eventos retornada com sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Criar um novo evento",
                "description": "Cria um novo evento e, opcionalmente, associa usuários ao evento.",
                "operationId": "dd683f690eaabf94d57d985cdf4f2120",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Evento de Exemplo"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-09-01"
                                    },
                                    "user_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Evento criado com sucesso"
                    },
                    "400": {
                        "description": "Erro na criação do evento"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                }
            }
        },
        "/api/v1/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Obter um evento específico",
                "description": "Retorna um evento pelo ID com os usuários associados.",
                "operationId": "cab1311cf1b9c45ea4813e7b5d0fd908",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do evento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evento retornado com sucesso"
                    },
                    "404": {
                        "description": "Evento não encontrado"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Atualizar um evento",
                "description": "Atualiza um evento existente pelo ID e, opcionalmente, sincroniza os usuários associados.",
                "operationId": "f2f9aa7bdb3bb955ea94ff859a7227d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do evento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Evento Atualizado"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-10-01T14:30:00Z"
                                    },
                                    "user_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Evento atualizado com sucesso"
                    },
                    "400": {
                        "description": "Erro na atualização do evento"
                    },
                    "404": {
                        "description": "Evento não encontrado"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Excluir um evento",
                "description": "Exclui um evento existente pelo ID.",
                "operationId": "864e3bd26fe09c0fecb665af777c091a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do evento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evento excluído com sucesso"
                    },
                    "404": {
                        "description": "Evento não encontrado"
                    },
                    "400": {
                        "description": "Erro ao excluir o evento"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Events",
            "description": "Events"
        }
    ]
}